+++ test/take.test.ts
@@ -1,6 +1,6 @@
 import { assertEquals } from "assert";
 
-import { asyncify, asyncIterableToArray, take } from "../mod.ts";
+import { arrayFrom, asyncify, take } from "../mod.ts";
 
 Deno.test("returns the first n elements from an iterable", async () => {
   const everyNumberSync = {
@@ -13,10 +13,10 @@ Deno.test("returns the first n elements from an iterable", async () => {
   assertEquals([...take(everyNumberSync, 7)], [0, 1, 2, 3, 4, 5, 6]);
 
   const everyNumberAsync = asyncify(everyNumberSync);
-  assertEquals(await asyncIterableToArray(take(everyNumberAsync, 0)), []);
-  assertEquals(await asyncIterableToArray(take(everyNumberAsync, 1)), [0]);
+  assertEquals(await arrayFrom(take(everyNumberAsync, 0)), []);
+  assertEquals(await arrayFrom(take(everyNumberAsync, 1)), [0]);
   assertEquals(
-    await asyncIterableToArray(take(everyNumberAsync, 4)),
+    await arrayFrom(take(everyNumberAsync, 4)),
     [0, 1, 2, 3],
   );
 });
@@ -51,11 +51,11 @@ Deno.test("iterates the minimum possible amount", async () => {
   }
 
   const async1 = new NumbersAsync();
-  await asyncIterableToArray(take(async1, 0));
+  await arrayFrom(take(async1, 0));
   assertEquals(async1.iterationCount, 0);
 
   const async2 = new NumbersAsync();
-  await asyncIterableToArray(take(async2, 3));
+  await arrayFrom(take(async2, 3));
   assertEquals(async2.iterationCount, 3);
 });
 
@@ -66,7 +66,7 @@ Deno.test("stops after the iterable has been exhausted", async () => {
   assertEquals([...take(set, 10000)], [1, 2, 3]);
 
   assertEquals(
-    await asyncIterableToArray(take(asyncify([1, 2, 3]), 3)),
+    await arrayFrom(take(asyncify([1, 2, 3]), 3)),
     [1, 2, 3],
   );
 });