+++ CHANGELOG.md
@@ -25,6 +25,8 @@
 - `at` should now be slightly faster for some types
 - **Breaking:** `range` now works like Python's `range`
 - **Breaking:** `zip` now stops after the shortest iterable
+- **Breaking:** `asyncIterableToArray` renamed to `arrayFrom`, now works with
+  sync iterables
 
 ### Removed